+++ .gitignore
@@ -52,3 +52,4 @@ out/
 .vscode/
 /worker/src/main/resources/firebase-key.json
 /fcm-test-frontend/
+/docs/
