+++ worker/src/main/java/com/mailsangja/worker/config/google/GoogleMailMessageClientConfig.java
@@ -1,19 +1,19 @@
-package com.mailsangja.worker.config;
+package com.mailsangja.worker.config.google;
 
-import com.mailsangja.worker.config.properties.GoogleMailWatchProperties;
+import com.mailsangja.worker.config.properties.GoogleMailInitialSyncProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
 
 @Configuration
-public class GoogleMailWatchClientConfig {
+public class GoogleMailMessageClientConfig {
 
     @Bean
-    public RestClient googleMailWatchRestClient(GoogleMailWatchProperties googleMailWatchProperties) {
+    public RestClient googleMailMessageRestClient(GoogleMailInitialSyncProperties properties) {
         SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-        requestFactory.setConnectTimeout((int) googleMailWatchProperties.getConnectTimeout().toMillis());
-        requestFactory.setReadTimeout((int) googleMailWatchProperties.getReadTimeout().toMillis());
+        requestFactory.setConnectTimeout((int) properties.getConnectTimeout().toMillis());
+        requestFactory.setReadTimeout((int) properties.getReadTimeout().toMillis());
 
         return RestClient.builder()
                 .requestFactory(requestFactory)
