+++ worker/src/main/java/com/mailsangja/worker/config/google/GoogleMailWatchClientConfig.java
@@ -1,19 +1,19 @@
-package com.mailsangja.worker.config;
+package com.mailsangja.worker.config.google;
 
-import com.mailsangja.worker.config.properties.GoogleOAuthProperties;
+import com.mailsangja.worker.config.properties.GoogleMailWatchProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
 
 @Configuration
-public class GoogleOAuthClientConfig {
+public class GoogleMailWatchClientConfig {
 
     @Bean
-    public RestClient googleOAuthRestClient(GoogleOAuthProperties googleOAuthProperties) {
+    public RestClient googleMailWatchRestClient(GoogleMailWatchProperties googleMailWatchProperties) {
         SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-        requestFactory.setConnectTimeout((int) googleOAuthProperties.getConnectTimeout().toMillis());
-        requestFactory.setReadTimeout((int) googleOAuthProperties.getReadTimeout().toMillis());
+        requestFactory.setConnectTimeout((int) googleMailWatchProperties.getConnectTimeout().toMillis());
+        requestFactory.setReadTimeout((int) googleMailWatchProperties.getReadTimeout().toMillis());
 
         return RestClient.builder()
                 .requestFactory(requestFactory)
