+++ worker/src/main/java/com/mailsangja/worker/config/google/GoogleOAuthClientConfig.java
@@ -1,19 +1,19 @@
-package com.mailsangja.worker.config;
+package com.mailsangja.worker.config.google;
 
-import com.mailsangja.worker.config.properties.GoogleMailInitialSyncProperties;
+import com.mailsangja.worker.config.properties.GoogleOAuthProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
 
 @Configuration
-public class GoogleMailMessageClientConfig {
+public class GoogleOAuthClientConfig {
 
     @Bean
-    public RestClient googleMailMessageRestClient(GoogleMailInitialSyncProperties properties) {
+    public RestClient googleOAuthRestClient(GoogleOAuthProperties googleOAuthProperties) {
         SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-        requestFactory.setConnectTimeout((int) properties.getConnectTimeout().toMillis());
-        requestFactory.setReadTimeout((int) properties.getReadTimeout().toMillis());
+        requestFactory.setConnectTimeout((int) googleOAuthProperties.getConnectTimeout().toMillis());
+        requestFactory.setReadTimeout((int) googleOAuthProperties.getReadTimeout().toMillis());
 
         return RestClient.builder()
                 .requestFactory(requestFactory)
